+++ crates/app/src/commands/daemon/server.rs
@@ -7,8 +7,8 @@ pub async fn server(session: MoonSession) -> AppResult {
     start_daemon_server(
         DaemonState {
             app_context: session.get_app_context().await?,
-            // Loaded in the background with the workspace watcher,
-            // otherwise it causes this command to block
+            // Loaded in the background within the workspace watcher,
+            // otherwise it causes this command to block for too long
             workspace_graph: Default::default(),
         },
         vec![Box::new(WorkspaceWatcher::new(session))],